Floatify Look

+++ module/planning/PlanLook/src/PlanLook.hpp
@@ -15,7 +15,7 @@ namespace module::planning {
             /// @brief How long to look at each search position
             float search_fixation_time = 0.0f;
             /// @brief List of search positions for the search pattern
-            std::vector<Eigen::Vector2d> search_positions{};
+            std::vector<Eigen::Vector2f> search_positions{};
         } cfg;
 
         /// @brief  Time since last search position transition
